Removed unnecessary type cast

+++ src/DotNext/RandomExtensions.cs
@@ -42,7 +42,7 @@ public static class RandomExtensions
         ref var firstChar = ref MemoryMarshal.GetReference(allowedChars);
         foreach (ref var element in buffer)
         {
-            var randomNumber = (uint)BitConverter.ToInt32(bytes.Span.Slice(offset)) % (uint)allowedChars.Length;
+            var randomNumber = BitConverter.ToUInt32(bytes.Span.Slice(offset)) % (uint)allowedChars.Length;
             element = Unsafe.Add(ref firstChar, randomNumber);
             offset += sizeof(int);
         }
